Create `quotes.cpp` and `quotes.h` files
